changing value for item attribute in model

+++ source/items_api/models.py
@@ -10,7 +10,7 @@ class Item(db.Model):
     __tablename__ = 'items'
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(80), unique=True, nullable=False)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
     description = db.Column(db.String(200), nullable=False)
     effect = db.Column(db.Float, nullable=False)
     sprite=db.Column(db.String(200), nullable=False)
